requirements.txt
python-telegram-bot==21.4
requests==2.32.3
APScheduler==3.10.4
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.30.1
